Copy fetchLinkSuggestions over to edit-navigation
