+++ packages/edit-navigation/package.json
@@ -22,6 +22,7 @@
 	"react-native": "src/index",
 	"dependencies": {
 		"@babel/runtime": "^7.9.2",
+		"@wordpress/api-fetch": "file:../api-fetch",
 		"@wordpress/block-editor": "file:../block-editor",
 		"@wordpress/block-library": "file:../block-library",
 		"@wordpress/blocks": "file:../blocks",
@@ -30,13 +31,14 @@
 		"@wordpress/data": "file:../data",
 		"@wordpress/data-controls": "file:../data-controls",
 		"@wordpress/dom-ready": "file:../dom-ready",
-		"@wordpress/editor": "file:../editor",
 	  	"@wordpress/element": "file:../element",
 		"@wordpress/hooks": "file:../hooks",
-		"@wordpress/i18n": "file:../i18n",
+	  	"@wordpress/html-entities": "file:../html-entities",
+	  	"@wordpress/i18n": "file:../i18n",
 		"@wordpress/keyboard-shortcuts": "file:../keyboard-shortcuts",
 		"@wordpress/media-utils": "file:../media-utils",
 		"@wordpress/notices": "file:../notices",
+		"@wordpress/url": "file:../url",
 		"classnames": "^2.2.5",
 		"lodash": "^4.17.15",
 		"rememo": "^3.0.0"
